Add generic implementation for singly-linked list
Implemented a generic singly-linked list that can hold values of any type using Go's new `any` keyword. The code includes a `Size` method that calculates the number of nodes in the list, and a `String` method to print out the values of all the nodes in the list. An example program is included in `main` function to illustrate how the list can be used.